+++ src/briefy/common/__init__.py
@@ -1,4 +1,5 @@
 """Briefy Common."""
 import translationstring
 
+
 _ = translationstring.TranslationStringFactory(__name__)