+++ tests/db/mixins/test_timestamp.py
@@ -16,7 +16,7 @@ class TimestampExample(Mixin, Base):
     __session__ = DBSession
 
 
-@pytest.mark.usefixtures("db_transaction")
+@pytest.mark.usefixtures('db_transaction')
 class TestTimestamp:
     """Test Timestamp database model."""